# swagger.yaml
openapi: 3.0.1
info:
  title: Exoplanet Classifier API
  description: API REST para clasificación automática de exoplanetas usando HistGradientBoostingClassifier. Permite entrenar modelos, realizar predicciones individuales y batch, y gestionar versiones de modelos.
  version: "1.0.0"
servers:
  - url: http://localhost:8000
    description: Servidor local de desarrollo
  - url: https://api.exoclassifier.com
    description: Servidor de producción
...
paths:
...
  /predict:
    post:
      tags: [Predict]
      summary: Predicción individual de exoplanetas
      description: Realiza predicciones individuales para uno o más exoplanetas usando una versión específica del modelo.
      parameters:
        - name: model_name
          in: query
          required: false
          schema:
            type: string
            default: "hgb_exoplanet_model"
          description: Nombre del modelo a usar
        - name: version
          in: query
          required: false
          schema:
            type: string
            default: "latest"
          description: Versión específica del modelo o 'latest'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                data:
                  type: array
                  items:
                    type: object
                    properties:
                      koi_period:
                        type: number
                        example: 10.5
                      koi_duration:
                        type: number
                        example: 2.1
                      koi_depth:
                        type: number
                        example: 0.001
                      koi_prad:
                        type: number
                        example: 1.2
                      koi_steff:
                        type: number
                        example: 5778
                      koi_slogg:
                        type: number
                        example: 4.4
      responses:
        "200":
          description: Predicciones exitosas
          content:
            application/json:
              schema:
                type: object
                properties:
                  predictions:
                    type: array
                    items:
                      type: object
                      properties:
                        class:
                          type: string
                          example: "CANDIDATE"
                        probabilities:
                          type: object
                          properties:
                            CANDIDATE:
                              type: number
                              example: 0.9244
                            CONFIRMED:
                              type: number
                              example: 0.0642
                            FALSE POSITIVE:
                              type: number
                              example: 0.0114
                  model_info:
                    type: object
                    properties:
                      model_name:
                        type: string
                        example: "hgb_exoplanet_model"
                      version:
                        type: string
                        example: "v1.0.0"
                      used_model:
                        type: string
                        example: "hgb_exoplanet_model:v1.0.0"
        "404":
          description: Versión del modelo no encontrada
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Versión 'v999.0.0' no encontrada para el modelo 'hgb_exoplanet_model'. Versiones disponibles: ['v1.0.0', 'v1.0.1', 'v1.0.2']"
  /predict/upload:
    post:
      tags: [Predict]
      summary: Predicción batch via archivo CSV
      description: Realiza predicciones batch subiendo un archivo CSV con datos de exoplanetas usando una versión específica del modelo.
      parameters:
        - name: model_name
          in: query
          required: false
          schema:
            type: string
            default: "hgb_exoplanet_model"
          description: Nombre del modelo a usar
        - name: version
          in: query
          required: false
          schema:
            type: string
            default: "latest"
          description: Versión específica del modelo o 'latest'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Archivo CSV con datos de exoplanetas
      responses:
        "200":
          description: Predicciones batch exitosas
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_planets:
                    type: integer
                    example: 2
                  class_distribution:
                    type: object
                    properties:
                      CANDIDATE:
                        type: integer
                        example: 2
                  download_url:
                    type: string
                    example: "/download/test_data_predictions_v1.0.0.csv"
                  model_info:
                    type: object
                    properties:
                      model_name:
                        type: string
                        example: "hgb_exoplanet_model"
                      version:
                        type: string
                        example: "v1.0.0"
                      used_model:
                        type: string
                        example: "hgb_exoplanet_model:v1.0.0"
        "404":
          description: Versión del modelo no encontrada
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Versión 'v999.0.0' no encontrada para el modelo 'hgb_exoplanet_model'. Versiones disponibles: ['v1.0.0', 'v1.0.1', 'v1.0.2']"
  /download/{filename}:
    get:
      tags: [Predict]
      summary: Descargar archivo de predicciones
      description: Descarga un archivo de predicciones generado por el endpoint /predict/upload.
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
          description: Nombre del archivo a descargar
      responses:
        "200":
          description: Archivo CSV con predicciones
          content:
            text/csv:
              schema:
                type: string
                format: binary
        "404":
          description: Archivo no encontrado
  /train:
    post:
      tags: [Train]
      summary: Reentrenar modelo con nuevos hiperparámetros
      description: Reentrena el modelo con nuevos hiperparámetros y crea una nueva versión.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                learning_rate:
                  type: number
                  example: 0.1
                  description: Tasa de aprendizaje
                max_leaf_nodes:
                  type: integer
                  example: 50
                  description: Número máximo de nodos hoja
                min_samples_leaf:
                  type: integer
                  example: 20
                  description: Mínimo de muestras por hoja
                early_stopping:
                  type: boolean
                  example: true
                  description: Habilitar parada temprana
      responses:
        "200":
          description: Entrenamiento exitoso
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "completed"
                  model_version:
                    type: string
                    example: "v1.0.2"
                  used_params:
                    type: object
                    properties:
                      learning_rate:
                        type: number
                        example: 0.1
                      max_leaf_nodes:
                        type: integer
                        example: 50
                      min_samples_leaf:
                        type: integer
                        example: 20
                      early_stopping:
                        type: boolean
                        example: true
  /model-info/{model_name}:
    get:
      tags: [Model Info]
      summary: Información detallada de un modelo específico
      description: Obtiene información detallada de un modelo específico incluyendo métricas y matriz de confusión.
      parameters:
        - name: model_name
          in: path
          required: true
          schema:
            type: string
          description: Nombre del modelo a consultar
          example: "hgb_exoplanet_model"
      responses:
        "200":
          description: Información del modelo
          content:
            application/json:
              schema:
                type: object
                properties:
                  model_name:
                    type: string
                    example: "hgb_exoplanet_model"
                  version:
                    type: string
                    example: "v1.0.1"
                  model_path:
                    type: string
                    example: "/path/to/model.pkl"
                  metrics:
                    type: object
                    properties:
                      accuracy:
                        type: number
                        example: 0.9352
                      CANDIDATE:
                        type: object
                        properties:
                          precision:
                            type: number
                            example: 0.8630
                          recall:
                            type: number
                            example: 0.8414
                          f1-score:
                            type: number
                            example: 0.8521
                          support:
                            type: number
                            example: 599.0
                  confusion_matrix:
                    type: array
                    items:
                      type: array
                      items:
                        type: integer
                    example: [[504, 91, 4], [63, 766, 4], [17, 2, 1418]]
                  files:
                    type: object
                    properties:
                      metrics:
                        type: string
                        example: "/path/to/metrics.json"
                      matrix:
                        type: string
                        example: "/path/to/matrix.npy"
        "404":
          description: Modelo no encontrado
  /model-versions/{model_name}:
    get:
      tags: [Model Versions]
      summary: Listar todas las versiones de un modelo
      description: Obtiene todas las versiones disponibles de un modelo específico.
      parameters:
        - name: model_name
          in: path
          required: true
          schema:
            type: string
          description: Nombre del modelo a consultar
          example: "hgb_exoplanet_model"
      responses:
        "200":
          description: Lista de versiones del modelo
          content:
            application/json:
              schema:
                type: object
                properties:
                  model_name:
                    type: string
                    example: "hgb_exoplanet_model"
                  versions:
                    type: array
                    items: { type: string }
                    example: ["v1.0.0", "v1.0.1"]
                  total_versions:
                    type: integer
                    example: 2
                  latest_version:
                    type: string
                    example: "v1.0.1"
        "404":
          description: Modelo no encontrado
...